+++ src/index.js
@@ -24,6 +24,7 @@ export { default as Tag } from './components/tag';
 export { default as Tooltip } from './components/tooltip';
 export { default as Notification } from './components/notification';
 export { default as Loading } from './components/loading';
+export { default as Collapse } from './components/collapse';
 export { PropTypes } from '../libs';
 export { Transition } from '../libs';
 export { View } from '../libs';